Updated `User` and related entities to use `int` as primary key type instead of `Long`. Adjusted repositories, services, and controllers accordingly. Added detailed CRUD operations in `UserController` with proper HTTP responses.

+++ server/src/main/java/com/assessment/tracker/server/persistence/domain/Account_Role.java
@@ -9,7 +9,7 @@ public class Account_Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    public Long acc_roleID;
+    public int acc_roleID;
 
     private Role role;
